feat!: typed tick rows across all bindings
Tick rows now expose logical types instead of wire encodings, uniformly across Rust, Python, TypeScript, C, C++, Arrow/pandas/polars columns, the CLI renderer, and the HTTP server.

The option right is the logical character everywhere: char in Rust ('C' / 'P', '\0' when contract identity is absent), a one-character string in Python and TypeScript, and a uint32_t Unicode scalar value in the C header occupying the same 4-byte slot the raw ASCII integer held. Both wire encodings (Number ASCII code, Text CALL/C/PUT/P) decode at the boundary; the integer no longer appears on any surface, including OptionContract rows and the CLI table output.

EodTick's time pair carries the vendor's v3 semantics under unit-suffixed names: created_ms_of_day (report creation, not a trade time) and last_trade_ms_of_day (the day's final trade; 0 on no-trade days, whose price columns are documented zero-fill). The schema columns bind the v3 wire headers (created / last_trade, capture-verified) directly, and the server emits the vendor's own JSON keys with contract expiration as an ISO YYYY-MM-DD string per the v3 response shape.

CalendarDay types its fields: is_open is a boolean and status carries the vendor day-type vocabulary (the tdbe CalendarStatus enum in Rust, the open / early_close / full_close / weekend strings in dynamic bindings, Arrow output, and the server, and an int32_t code plus tdx_calendar_status_name() in C). A rows-present calendar response missing the type column now fails with a typed missing-header error instead of silently filling closed days.

Absent contract identity is None / undefined / Arrow null in Python and TypeScript, matching the streaming payload convention; the repr(C) Rust and C rows keep their documented fills with has_contract_id() as the presence check. The generators route every emitted field through per-language reserved-word tables, so the Python keyword collision on the Greeks lambda column is rendered lambda_ (attribute, constructor kwarg, repr) while columnar outputs keep the logical name, and a future keyword-colliding column cannot produce an unreachable attribute in any binding.

Rows carrying date plus a milliseconds-of-day column gain read-side epoch accessors (timestamp_ms, created_timestamp_ms, last_trade_timestamp_ms, underlying_timestamp_ms, quote_timestamp_ms) in Rust and Python, tdx_timestamp_ms in C, and tdx::timestamp_ms in C++; raw integer fields stay primary. List endpoints return values sorted ascending, numeric-aware for strike and date lists. DataFrame and Arrow column names follow the public field names (symbol, date) instead of wire spellings. The null-fill on bulk column extraction is the generator-emitted seed literal rather than a Default-trait obligation, so logical enum columns need no artificial Default.

+++ ffi/src/types.rs
@@ -278,9 +278,14 @@ tick_array_free!(tdx_trade_quote_tick_array_free, TdxTradeQuoteTickArray);
 pub struct TdxOptionContract {
     /// Heap-allocated NUL-terminated C string. Freed with the array.
     pub symbol: *const c_char,
+    /// Expiration date as a `YYYYMMDD` integer.
     pub expiration: i32,
+    /// Strike price in dollars.
     pub strike: f64,
-    pub right: i32,
+    /// Contract right as the Unicode scalar value of the character:
+    /// `'C'` (67) for a call, `'P'` (80) for a put. Cast to `char` for
+    /// display. Same 4-byte slot the previous ASCII integer occupied.
+    pub right: u32,
 }
 
 /// Array of FFI-safe option contracts.
@@ -308,7 +313,7 @@ impl TdxOptionContractArray {
                     symbol: CString::new(c.symbol)?.into_raw().cast_const(),
                     expiration: c.expiration,
                     strike: c.strike,
-                    right: c.right,
+                    right: c.right as u32,
                 })
             })
             .collect::<Result<Vec<_>, std::ffi::NulError>>()?;
